move complex handleInitialValue logic to own file
