+++ packages/web-components/src/components/rux-datetime-picker/rux-datetime-picker.tsx
@@ -27,6 +27,7 @@ import {
     toggleCalendar,
     highlightInput,
 } from './datetime-picker.handlers'
+import { handleInitialValue } from './datetime-picker.value-processing'
 import {
     getMaskClasses,
     getDisplayClasses,
@@ -34,18 +35,9 @@ import {
     getInputContainerClasses,
 } from './datetime-picker.render'
 import {
-    buildMicroOrdinalIsoString,
     combineToISO,
     formatOrdinalToIso,
-    getMonthFromDayOfYear,
-    getMonthValueByName,
-    initialOrdinalParts,
-    initialParts,
     isIsoString,
-    isLeapYear,
-    julianToGregorianDay,
-    setIsoPart,
-    setJulianIsoPart,
     setMaxLength,
     setMaxLengthOrdinal,
     setPart,
@@ -53,8 +45,6 @@ import {
     toPartialOrdinalIsoString,
     toPartialRegularIsoString,
 } from './utils'
-
-import { buildMicroIsoString } from './utils/index'
 import { renderHiddenInput } from '../../utils/utils'
 
 @Component({
@@ -282,283 +272,15 @@ export class RuxDatetimePicker
     }
 
     handleInitialValue(value?: string) {
-        const initial = this.julianFormat
-            ? initialOrdinalParts()
-            : initialParts()
-        const isMicro = this.precision === 'us'
-        if (value) {
-            try {
-                if (this.julianFormat && value.length === 3) {
-                    const currentYear = new Date().getUTCFullYear()
-                    value = `${currentYear}-${value}`
-                }
-
-                // --- MICROSECOND HANDLING ---
-                if (isMicro && !this.julianFormat) {
-                    //check if incoming value is in an Oridnal ISO format. If so, convert it to gregorian since we're not in julianFormat.
-                    const ordinalFormatMatch = value.match(
-                        /^([0-9]{4})(?:-([0-9]{1,3}))?(?:T([0-9]{2})(?::([0-9]{2}))?(?::([0-9]{2})(?:\.([0-9]{1,6}))?)?Z?)?$/
-                    )
-                    if (ordinalFormatMatch) {
-                        // Extract year, ordinal day, hour, min, sec, micro
-                        const year = ordinalFormatMatch[1] || '0000'
-                        const ordinal = ordinalFormatMatch[2] || '001'
-                        const hour = ordinalFormatMatch[3] || '00'
-                        const min = ordinalFormatMatch[4] || '00'
-                        const sec = ordinalFormatMatch[5] || '00'
-                        const micro = ordinalFormatMatch[6] || '000000'
-
-                        // Convert ordinal day to month and day
-                        const date = new Date(Date.UTC(Number(year), 0, 1))
-                        date.setUTCDate(Number(ordinal))
-                        const month = (date.getUTCMonth() + 1)
-                            .toString()
-                            .padStart(2, '0')
-                        const day = date
-                            .getUTCDate()
-                            .toString()
-                            .padStart(2, '0')
-
-                        value = buildMicroIsoString({
-                            year,
-                            month,
-                            day,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        })
-                    }
-                    // Try to extract all parts from the value (ISO or partial ISO)
-                    // Accepts both YYYY-MM-DDTHH:mm:ss.SSSSSSZ and partials
-
-                    const regex = /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2})\.(\d{1,6})Z?$/
-                    const match = value.match(regex)
-                    let iso = ''
-                    if (match) {
-                        const [
-                            ,
-                            year,
-                            month,
-                            day,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        ] = match
-                        iso = buildMicroIsoString({
-                            year,
-                            month,
-                            day,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        })
-                    } else {
-                        // Fallback: try to parse as much as possible
-                        const partial = value.match(
-                            new RegExp(
-                                `^(\d{4})-(\d{2})-(\d{2})(?:T(\d{2}))?(?::(\d{2}))?(?::(\d{2}))?(?:\.(\d{1,6}))?Z?$`
-                            )
-                        )
-                        const [
-                            ,
-                            year = '0000',
-                            month = '01',
-                            day = '01',
-                            hour = '00',
-                            min = '00',
-                            sec = '00',
-                            micro = '000000',
-                        ] = partial || []
-                        iso = buildMicroIsoString({
-                            year,
-                            month,
-                            day,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        })
-                    }
-                    // Set initial part values from ISO string
-                    for (const part of initial) {
-                        if (part.type === 'mask') continue
-                        part.value = setIsoPart[part.type](iso, true)
-                    }
-                    this.iso = iso
-                    this.value = iso
-                    this.parts = initial
-                    return
-                }
-                if (this.julianFormat && isMicro) {
-                    const match = value.match(
-                        /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2})\.(\d{1,6})Z$/
-                    )
-                    if (match) {
-                        const [
-                            ,
-                            year,
-                            month,
-                            day,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        ] = match
-                        // Calculate ordinal day
-                        const date = new Date(
-                            `${year}-${month}-${day}T00:00:00Z`
-                        )
-                        const start = new Date(Date.UTC(Number(year), 0, 0))
-                        const ordinalDay = String(
-                            Math.floor(
-                                (date.getTime() - start.getTime()) /
-                                    (1000 * 60 * 60 * 24)
-                            )
-                        ).padStart(3, '0')
-                        value = buildMicroOrdinalIsoString({
-                            year,
-                            jday: ordinalDay,
-                            hour,
-                            min,
-                            sec,
-                            micro,
-                        })
-                    }
-                }
-                // --- END MICROSECOND HANDLING ---
-
-                const ordinalFormatMatch = value.match(
-                    /^([0-9]{4})(?:-([0-9]{1,3}))?(?:T([0-9]{2})(?::([0-9]{2}))?(?::([0-9]{2})(?:\.([0-9]{1,6}))?)?Z?)?$/
-                )
-
-                let d: Date | undefined | string = undefined
-
-                if (ordinalFormatMatch && this.julianFormat) {
-                    // Parse year and day-of-year
-                    const year =
-                        ordinalFormatMatch[1] ||
-                        new Date().getUTCFullYear().toString()
-                    let jdayNum = parseInt(ordinalFormatMatch[2] || '1', 10)
-                    if (isNaN(jdayNum) || jdayNum < 1) jdayNum = 1
-
-                    const yearNum = parseInt(year, 10)
-                    const maxDay = isLeapYear(yearNum) ? 366 : 365
-                    if (jdayNum > maxDay) jdayNum = maxDay
-                    const jday = jdayNum.toString().padStart(3, '0')
-                    const hour = ordinalFormatMatch[3] || '00'
-                    const minute = ordinalFormatMatch[4] || '00'
-                    const sec = ordinalFormatMatch[5] || '00'
-                    const ms = ordinalFormatMatch[6]
-                        ? ordinalFormatMatch[6]
-                        : isMicro
-                        ? '000000'
-                        : '000'
-                    const gregDay = julianToGregorianDay(jday, year).padStart(
-                        2,
-                        '0'
-                    )
-                    const month = getMonthValueByName(
-                        getMonthFromDayOfYear(jday, yearNum)!
-                    )
-                    if (!isMicro) {
-                        d = new Date(
-                            `${year}-${month}-${gregDay}T${hour}:${minute}:${sec}.${ms}Z`
-                        )
-                    } else {
-                        d = buildMicroOrdinalIsoString({
-                            year,
-                            jday,
-                            hour,
-                            min: minute,
-                            sec,
-                            micro: ms,
-                        })
-                    }
-                } else {
-                    // Special case: 2-digit value as month (01-12)
-                    if (
-                        value.length === 2 &&
-                        /^\d{2}$/.test(value) &&
-                        parseInt(value, 10) >= 1 &&
-                        parseInt(value, 10) <= 12
-                    ) {
-                        // Treat as month, default year to current year
-                        const currentYear = new Date().getUTCFullYear()
-                        d = new Date(`${currentYear}-${value}`)
-                    } else {
-                        // Try to parse as a direct date first
-                        d = new Date(value)
-                        if (isNaN(d.getTime())) {
-                            // Fallback: try to parse as partial ISO
-                            const iso = toPartialRegularIsoString(value)
-                            d = new Date(iso)
-                        }
-                    }
-                }
-
-                // if (!d || isNaN(d.getTime())) throw new Error('Invalid date')
-
-                // Always get ISO string (Gregorian)
-                let iso
-                if (d instanceof Date) {
-                    iso = d.toISOString()
-                    if (this.julianFormat) {
-                        iso = toOrdinalIsoString(iso)
-                    }
-                } else {
-                    iso = d
-                }
-
-                // Set initial part values from ISO string
-                for (const part of initial) {
-                    if (part.type === 'mask') continue
-                    part.value = this.julianFormat
-                        ? setJulianIsoPart[part.type](iso, isMicro)
-                        : setIsoPart[part.type](iso, isMicro)
-                }
-
-                // Always pass down Gregorian ISO to calendar
-                this.iso = formatOrdinalToIso(iso)
-            } catch (error: any) {
-                this.iso = error.message || 'Invalid date'
-            }
-        }
-
-        // Adjust parts array length based on precision
-        switch (this.precision) {
-            case 'day':
-                // Remove all time parts for day precision
-                !this.julianFormat ? initial.splice(5, 9) : initial.splice(3, 9)
-                break
-            case 'hour':
-                // Keep only up to hour
-                !this.julianFormat ? initial.splice(7, 6) : initial.splice(5, 6)
-                break
-            case 'min':
-                !this.julianFormat ? initial.splice(9, 4) : initial.splice(7, 4)
-                break
-            case 'sec':
-                !this.julianFormat
-                    ? initial.splice(11, 2)
-                    : initial.splice(9, 2)
-                break
-            case 'ms':
-                break
-            case 'us':
-                break
-            default:
-                initial.splice(9, 4)
-                break
-        }
-
-        this.parts = initial
+        const result = handleInitialValue(
+            value,
+            this.julianFormat,
+            this.precision
+        )
+        this.iso = result.iso
+        this.parts = result.parts
+        this.value = result.finalValue
         console.log('this.iso Before this.value assignment: ', this.iso)
-        this.value = this.julianFormat
-            ? toPartialOrdinalIsoString(this.iso, isMicro)
-            : this.iso
         console.log('End of HIV value: ', this.value)
     }
